+++ DebugSwift/Sources/Features/Network/Persistence/NetworkSessionPersistenceManager.swift
@@ -11,107 +11,6 @@ import Foundation
 import SwiftData
 
 @available(iOS 17.0, *)
-@Model
-final class NetworkSessionEntity {
-    @Attribute(.unique) var id: UUID
-    var startedAt: Date
-    var endedAt: Date?
-    var createdAt: Date
-    @Relationship(deleteRule: .cascade, inverse: \NetworkRequestEntity.session)
-    var requests: [NetworkRequestEntity]
-
-    init(
-        id: UUID = UUID(),
-        startedAt: Date,
-        endedAt: Date? = nil,
-        createdAt: Date = Date(),
-        requests: [NetworkRequestEntity] = []
-    ) {
-        self.id = id
-        self.startedAt = startedAt
-        self.endedAt = endedAt
-        self.createdAt = createdAt
-        self.requests = requests
-    }
-}
-
-@available(iOS 17.0, *)
-@Model
-final class NetworkRequestEntity {
-    @Attribute(.unique) var id: UUID
-    var url: String?
-    var requestData: Data?
-    var responseData: Data?
-    var requestId: String?
-    var method: String?
-    var statusCode: String?
-    var mineType: String?
-    var startTime: String?
-    var endTime: String?
-    var totalDuration: String?
-    var isImage: Bool
-    var isEncrypted: Bool
-    var requestHeadersData: Data?
-    var responseHeadersData: Data?
-    var errorDescriptionText: String?
-    var errorLocalizedDescriptionText: String?
-    var size: String?
-    var modelIndex: Int
-    var mode: String
-    var capturedAt: Date
-    var createdAt: Date
-    var session: NetworkSessionEntity?
-
-    init(
-        id: UUID = UUID(),
-        url: String?,
-        requestData: Data?,
-        responseData: Data?,
-        requestId: String?,
-        method: String?,
-        statusCode: String?,
-        mineType: String?,
-        startTime: String?,
-        endTime: String?,
-        totalDuration: String?,
-        isImage: Bool,
-        isEncrypted: Bool,
-        requestHeadersData: Data?,
-        responseHeadersData: Data?,
-        errorDescriptionText: String?,
-        errorLocalizedDescriptionText: String?,
-        size: String?,
-        modelIndex: Int,
-        mode: String,
-        capturedAt: Date = Date(),
-        createdAt: Date = Date()
-    ) {
-        self.id = id
-        self.url = url
-        self.requestData = requestData
-        self.responseData = responseData
-        self.requestId = requestId
-        self.method = method
-        self.statusCode = statusCode
-        self.mineType = mineType
-        self.startTime = startTime
-        self.endTime = endTime
-        self.totalDuration = totalDuration
-        self.isImage = isImage
-        self.isEncrypted = isEncrypted
-        self.requestHeadersData = requestHeadersData
-        self.responseHeadersData = responseHeadersData
-        self.errorDescriptionText = errorDescriptionText
-        self.errorLocalizedDescriptionText = errorLocalizedDescriptionText
-        self.size = size
-        self.modelIndex = modelIndex
-        self.mode = mode
-        self.capturedAt = capturedAt
-        self.createdAt = createdAt
-    }
-}
-
-@available(iOS 17.0, *)
 @MainActor
 final class NetworkSessionPersistenceManager {
     private enum Preference {
@@ -404,32 +303,6 @@ final class NetworkSessionPersistenceManager {
         return headers.reduce(into: [String: String]()) { partialResult, item in
             partialResult[item.key.lowercased()] = item.value
         }
-    }
-}
-
-@available(iOS 17.0, *)
-extension NetworkRequestEntity {
-    func makeHttpModel() -> HttpModel {
-        let model = HttpModel()
-        model.url = url.flatMap(URL.init(string:))
-        model.requestData = requestData
-        model.responseData = responseData
-        model.requestId = requestId
-        model.method = method
-        model.statusCode = statusCode
-        model.mineType = mineType
-        model.startTime = startTime
-        model.endTime = endTime
-        model.totalDuration = totalDuration
-        model.isImage = isImage
-        model.isEncrypted = isEncrypted
-        model.requestHeaderFields = NetworkSessionPersistenceManager.decodeHeaders(requestHeadersData)
-        model.responseHeaderFields = NetworkSessionPersistenceManager.decodeHeaders(responseHeadersData)
-        model.errorDescription = errorDescriptionText
-        model.errorLocalizedDescription = errorLocalizedDescriptionText
-        model.size = size
-        model.index = modelIndex
-        return model
     }
 }
 #endif
